refactor(design-system/checkbox): checkbox 컴포넌트 수정 - 첫자 소문자로 변경 CheckboxGroup -> checkboxGroup
